+++ shoop/core/order_creator/_source.py
@@ -264,10 +264,13 @@ class OrderSource(object):
         return False
 
     def add_line(self, **kwargs):
-        line = SourceLine(source=self, **kwargs)
+        line = self.create_line(**kwargs)
         self._lines.append(line)
         self.uncache()
         return line
+
+    def create_line(self, **kwargs):
+        return SourceLine(source=self, **kwargs)
 
     def get_lines(self):
         """
@@ -285,9 +288,6 @@ class OrderSource(object):
         :rtype: decimal.Decimal|int
         """
         return sum([line.quantity for line in self.get_product_lines()])
-
-    def create_line(self, **kwargs):
-        return SourceLine(source=self, **kwargs)
 
     def get_final_lines(self, with_taxes=False):
         """
@@ -350,7 +350,7 @@ class OrderSource(object):
 
     def _compute_processed_lines(self):
         # This function would be a good candidate for subclass extension.
-        lines = self.get_lines()
+        lines = list(self.get_lines())
 
         lines.extend(self._compute_payment_method_lines())
         lines.extend(self._compute_shipping_method_lines())
@@ -502,14 +502,7 @@ class SourceLine(TaxableItem, Priceful):
         self.require_verification = kwargs.pop("require_verification", False)
         self.accounting_identifier = kwargs.pop("accounting_identifier", "")
 
-        self.taxes = []
-        """
-        Taxes of this line.
-
-        Determined by a TaxModule in :func:`OrderSource.calculate_taxes`.
-
-        :type: list[shoop.core.taxing.LineTax]
-        """
+        self._taxes = None
 
         self._data = kwargs.copy()
 
@@ -581,13 +574,37 @@ class SourceLine(TaxableItem, Priceful):
         self._tax_class = value
 
     @property
+    def taxes(self):
+        """
+        Taxes of this line.
+
+        Determined by a TaxModule in :func:`OrderSource.calculate_taxes`.
+
+        :rtype: list[shoop.core.taxing.LineTax]
+        """
+        return self._taxes or []
+
+    @taxes.setter
+    def taxes(self, value):
+        assert isinstance(value, list)
+        assert all(isinstance(x, taxing.LineTax) for x in value)
+        self._taxes = value
+
+    @property
     def tax_amount(self):
         """
         :rtype: shoop.utils.money.Money
         """
         self.source.calculate_taxes_or_raise()
         zero = self.source.zero_price.amount
-        return sum((x.amount for x in self.taxes), zero)
+        taxes = self._taxes
+        if taxes is None:
+            # Taxes were calculated to a different line instance.  Get
+            # taxes from there.
+            for line in self.source.get_final_lines():
+                if line.line_id == self.line_id and line.price == self.price:
+                    taxes = line.taxes
+        return sum((x.amount for x in taxes), zero)
 
     def _serialize_field(self, key):
         value = getattr(self, key)